pyproject.toml
[project]
name = "ttihp-spi-ws2812"
version = "0.1.0"
description = "Amaranth code for driving x16 WS2812 LED strips using a custom ASIC on TinyTapeout"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amaranth>=0.5.7",
    "amaranth-yosys>=0.50.0.0.post115",
    "atomics>=1.0.3",
    "cocotb>=1.9.2",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
]
